+++ src/config/commandConfig.js
@@ -781,6 +781,7 @@ const commandConfigRaw = {
           "steven",
           "palmer",
           "professorWillow",
+          "butler",
         ],
       },
       difficulty: {